[f5121] avoid incorrect usage of Android GPS HAL in certain cases. MER#1871
This adds a patch for adding the aapcs calling convention to the gps.h header.

By adding the correct calling convention to the gps.h headers, we avoid that calls to for
example inject_location form hybris-processes provide wrong data to the HAL.

inject_location is used to provide position data (i.e. acquired through online services) to
the Android GPS HAL. This means without this change, an invalid location would be sent to the
Android GPS HAL, possibly confusing it.